Use css to change the appearance of div, p, button
I used styles.css to change the background color,
font color, font size, and font weight of the div
element. I used the style element in index.html to
change the background color, font color, and font
size of the paragraph element. I used the style
attribute to change the background color and font
size of the button element.