Add classes and ids to elements for unique styling
The "odd-element" class was added to odd numbered
elements to style them together at once. An id was
added to the second element to apply unique styles
to it. The third element was given an additional
class to apply extra styles to it. The fourth
element was given an id to spply unique styles to
it.