Add padding and gap to the modal
Adding padding to the modal adds space between the
edges of the modal and every element inside of it.
Adding gap to the modal adds space between the
icon and the main content to the right of it.
Making the padding and gap the same value makes
the space around the icon the same.