Turn the header into a flex container
Turning the header in a flex container makes all
the links on the top of the page aligned
horizontally and allows me to add the appropriate
space in the middle of the 2 items. In addition,
padding was added to the header to make the links
easier to click and to make the page look more
similar to the desired outcome.